#1044 5. Reduced code duplication by extracting common numeric editor logic into tg-numeric-editor.
